feat: Add experimental new fast runtime
Uses an array instead of an object for storing refs. This prevents mutating the shadow class of the object used to store the refs for each ref. Looking up refs is now also done my array index which is faster than an object lookup. Combined it should make a significant difference when `collect` is used in a hot loop and/or with a large amount of refs in a template.